requirements.txt
# CoreML tools requires Python 3.9-3.11
# Install with: python3.11 -m venv venv && source venv/bin/activate && pip install -r requirements.txt

torch==2.4.0
torchvision==0.19.0
coremltools==8.0
numpy==1.26.4
huggingface_hub==0.26.0
gdown==4.7.1

# Additional dependencies for Real-ESRGAN upscaler
opencv-python==4.8.1.78
Pillow==10.4.0
gfpgan==1.3.8
basicsr==1.4.2
realesrgan==0.3.0